requirements.txt
fastapi 
uvicorn
langchain-cohere
python-dotenv 
psycopg2-binary